Drop LeaderChange; refactor PaceMaker
LeaderChange served mostly the same purpose as NewView, but with
added authentication. However, it should not be necessary to
authenticate a QC as the a valid QC itself is proof that it is
legitimate.

As an added bonus, dropping out leaderchange will allow separating
pacemaker out of the hotstuff package entirily in the future.