Remove block interface
We will no longer be implementing the various hotstuff interfaces on the
Protobuf types, since they conflict with the goals of the hotstuff interfaces in
some areas. For example, the protobuf types implement stringer, which
means that we can't write our own stringer implementation for these
types. Instead, we will just convert between the protobuf types and the
hotstuff types.

Because of this, I see no need to keep the Block interface around, which
simplifies things.