Implement block fetching
This adds the ability to fetch a missing block from the other replicas.
Currently, it is only used in cases where a replica receives a vote for
a block, but does not have the block locally. In this case, it may be
that the replica has received the vote before the block arrived, so we
make sure to abort the fetch if the block received in time.