Make it easier to reference other modules
This commit adds a new HotStuff struct that holds references to all
modular components used within HotStuff. It also adds a Module interface
that can be implemented in order to receive a pointer to the HotStuff
struct. The HotStuff struct and the modules it contains are initialized
using the Builder struct.

For example, the chainedhotstuff implementation needs to access other
modules such as the Signer/Verifier and Config modules. To gain access
to these, the chainedhotstuff struct implements the Module interface.
When registered with a Builder, the chainedhotstuff struct will be
stored in the HotStuff object's Consensus field, since chainedhotstuff
implements the Consensus interface. In addition, the Builder will call
the InitModule, which is part of the Module interface, on the
chainedhotstuff struct. This gives the chainedhotstuff struct a
reference to the HotStuff object, which has getters for the other
modules it needs.