Refactor block fetching
This makes use of the work-in-progress channels in Gorums
(see https://github.com/relab/gorums/pull/131).
We can use a dedicated channel for all messages that are core to the
protocol, such that they are all handled synchronously, and then another
channel for other messages, such as fetch requests/responses.
This allows us to do RPC messages on the event loop goroutine.
We can thus make the fetch operation part of the BlockChain.Get()
method, which simplifies things a lot.

However, logs show that allowing each call to Get() to attempt to fetch
blocks from other replicas will generate a lot of unnecessary fetch
attempts. To combat this, we add a LocalGet() method which should be
used instead of Get() when it is undesirable to fetch a missing block.

Special attention is needed in OnVote(), because it can very often fail
to find its blocks. Thus, we add the AwaitEvent() method to the event
loop. This method allows us to reschedule a vote event to run after
the next proposal event occurs. This allows us to avoid around 90% of
OnVote-initiated fetch attempts, according to my measurements.
OnVote will attempt to reschedule the event the first time it fails to
find its block. The second time, it will try to fetch the block.