actually we don't need channels to do fetching
I initially thought that invoking RPCs from the event loop would cause a
deadlock because the event loop would be waiting for the RPC to be
while the RPC was waiting to be handled by the event loop.
Then, I remembered that each replica has its own send/receive stream for
making requests to another replica. So if we have two replicas a and b,
then replica a has one stream for sending requests to b, and b has one
stream for sending requests to a. Now, if a sends a vote message to b,
then b can safely fetch the proposal from a because it would use the
other stream.