Allow commands to be late
This renames the GetCommand() method of CommandQueue to Get(),
and adds a context parameter. This allows the CommandQueue to wait until
it has a command or until the context is cancelled.

This also means that HotStuff will no longer propose empty commands.
Instead, we will wait until either the view expires, or a command is
available.