main_test: get free ports on each run
Hardcoding the ports is problematic because it can cause collisions.
This commit makes the test get free ports from the kernel each time.
Because we are getting the free ports by opening and closing listeners,
it could still be possible for collisions to occur. Hopefully this will
work for now.