hopefully fixed main_test
main_test has been consistently passing for me locally while
often failing on github actions. I think the reason is that some
replicas were unable to write the last commands to the output
file before the test stops the event loop. This commit attempts to fix
that by allowing the event loop to finish the events that were in the
event queue at the time when the context is cancelled. We also need to
ensure that main_test does not start processing the output files until
the event loops have stopped.