make it possible to accept a command twice if voting fails
If a proposal does not get the required amount of votes, it should still be
possible to propose the same commands again. To allow this, we add a new
Proposed() method to the acceptor interface, which tells the acceptor
that the proposal for that command was successfull, and the command
should not be accepted again in the future.