remove VotingMachine interface
This removes the VotingMachine interface,
and the VotingMachine getter from the modules object.
We no longer need the interface because if move VotingMachine
back into the consensus package. We also no longer need to be able to
access the VotingMachine through the Modules object because we can
instead send it events.