eventloop: add observers
This commit changes way handlers work and adds the concept of observers.
Previously, we could register multiple async handlers per event type,
but only one synchronous handler. Async handlers were allowed to consume
events, but it was not possible to ensure the order that async handlers
were rexecuted in. If we have two consumers A and B with async handlers.
If A consumes the event, then B would not receive the event unless it
was registered before A.

With this change, there can only be one handler per event type, and all
handlers consume events. To make an event visible to multiple consumers,
we can instead register observers. Observers are just like handlers,
except they do not consume the events. In addition, observers are
executed before any handlers.

Thus, we can solve the problem described earlier by making B an
observer. This guarantees that B will see the event before A consumes
it. Note that an async handler will still prevent synchronous observers
from observing an event.