add a simplified interface for implementing consensus
This adds a Rules interface to the consensus package.
The Rules interface specifies the rules for voting for and committing
blocks. Implementors of the Rules interface can use the consensus.New()
method to wrap a Rules interface with a struct that implements the
old Consensus interface.

Note that this does not change the old Consensus interface, instead it
adds a simplified interface that reduces the amount of code that has to
be written for new consensus protocols. The old interface may be useful
in cases where the Rules interface is too simplified.