Add a registry for modules
This adds a registry in the modules package that can be used to register
and retrieve modules by name and interface. Modules register by calling
RegisterModule, specifying a module name, and passing a constructor
function. Note that this constructor function's return value must be the
interface for the module. Then, a module can be retrieved by calling
GetModule with the module name and a pointer to a variable where the
module should be stored. The pointer must be a non-nil pointer to an
interface type.