twins: avoid waiting on synchronizer timeout
Instead we detect when timeouts should happen by manually ticking each
event loop in rounds and then monitoring whether a replica sends any
messages or not. A replica that is participating in a view will send a
message, and should not raise a timeout. A replica that sends no
messages should raise a timeout.