twins: add a scenarios per file option
This allows specifying a maximum amount of scenarios that can be written
to a single file, after which the generator will create a new file to
write to. This is useful for limiting the size of the generated files.
When this option is specified, the --output parameter is used
as a directory path instead of a file path.