CryptoBase: operate on messages instead of hashes
The old interface passed hashes of messages to the Sign and Verify
functions. With this changes, the messages should be passed in raw byte
form instead. The main reason for making this change is that the
bls12 implementation hashes the message as part of the HashToCurve
operation, so in the old versions, the messages were effectively hashed
twice.