synchronizer: create an internal updateHighQC
This commit introduces an internal method updateHighQC, which is the
same as the exported UpdateHighQC, except it does not verify the qc
first. This is useful for AdvanceView because it verifies the qc before
calling updateHighQC. Thus we, can save some resources by only verifying
the QC once. When caching is enabled, the savings will not be great,
but I still think it makes sense to do this.