consensus: Fix bug in commitInner
When the commitInner() method is going to commit a block, let's call it
b, it must also commit all uncommitted blocks between b and the last
committed block, let's call it bExec. It does this by recursively
calling commitInner() with the parent block until bExec is reached.

The problem is this: the BlockChain module's Get() method can fail to
locate the parent block. The code did not take this into account. If the
parent block could not be foundting the child.

For example, if we have the follo, the commitInner() method would move
on to commitwing branch:

bExec <- b' <- b

If we want to commit b, but the BlockChain module fails to find b',ecome
the new bExec. Hence, b' is never committed. then the code would still
commit b, and b would become the new bExec. Hence, b' is never
committed.

A fix was attempted in 591e5f63, however, I now realize it was not
sufficient. Consider the following:

bExec <- b'' <- b' <- b

If commitInner(b') fails to locate b'' and refuses to commit, then
commitInner(b) will still commit, because it does not know whether
commitInner(b') was successful or not. Therefore, it is necessary to
return an error and check it after every invocation of commitInner.