Implicit modules
This removes the need to declare what "interface" or type a module
provides to the module system, thus being more in line with the "Go way"
of implicit interfaces. The module system looks for a module with a
compatible type. If multiple modules implement the same type, the module
system chooses the one that was added to the builder last.