Added package comments that caused revive linter to bark
These are quick and dirty package comments and may not provide
a super accurate description. However, we may want to reconsider
the package structure of the project to avoid some generic
packages like the types package (in metrics)... We could just
keep the proto files in the metrics package instead.