modules: Make it possible to use custom options
The options struct is used for configuration that is shared between
modules. With this commit, we no longer need to add new options as a
field in the options struct, but allow packages to register new options
at will. The options are stored in an array and indexed using the
option ID, so the performance impact should be minimal.