Removed replicaLocationInfo map from Experiment struct
This avoids the location info map in the Experiment struct since we can populate
the replicaOpts.LocationInfo map from the assignReplicasAndClients method instead
of in the createReplicas method. We do this by reusing the location info map
across all the replicaOpts; I don't think this is a problem since the map should
not change after the assignReplicasAndClients has been called.