Clean up the proto to Go struct literal and more
This commit replaces the tostring.go file with the proto2.GoString()
function that takes a generic proto.Message and returns a Go struct
literal. This commit also replaces the interfaces AlmostFuzzMsg and
FuzzMsgInterface with two simple helpers extractProtoMsg() and
fuzzMsgToHotStuffMsg(). These are easier to use and IMO easier
to maintain. Additionally, there are a few smaller tweaks.