modules/registry: fix ListModules with interface types
This was broken since fc203fc9

The problem is that RegisterModule looks at the zero value of a module's
type. However, this does not work for interface types. A workaround is
to get the type of a pointer to the interface and then access its
element type.

Fixes --list-modules CLI flag produces segfault #101