synchronizer: fix broken timer
The timer used by the synchronizer used to be resettable, but after
7776a29, the closure passed to the timer should only be used once
because it stores the current view. This was done so that we can avoid
the need for any synchronization of access to the current view in the
synchronizer. However, I failed to realize that this would make the
timer only usable once. Thus, resetting it is not possible as the view
that is stored in the timer's closure would not change upon reset.

This commit wraps the timer in a `oneShotTimer` struct that is intended
to prevent the timer from being reused. Instead, the methods
`startTimeoutTimer` and `stopTimeoutTimer` should be used instead of
using the timer directly. These functions ensure that each timer is
unique and thus a unique closure is used each time, ensuring that we get
the correct view and avoid the need for synchronization code.