refactor(latencies): renamed LatencyCity and LatencyID to Latency
This renames the LatencyCity function that operates on the full
set of locations. This is now called Latency(a, b string).

This also renames LatencyID method on the LatencyMatrix to just
Latency(a, b hotstuff.ID) as well. It should not be possible to
confuse the two; one is a method on LatencyMatrix and takes a pair
of hotstuff.ID values, whereas the other is a function and takes
a pair of string location values.

This also improves documentation, explaining that providing invalid
input to the Latency functions will cause a panic.