refactor(backend): uses new streamlined latency.Matrix to add delay
This replaces the old multi-map-based location and latencies lookups
that were a bit awkward to work with.

This removes the existing map[string]map[string]time.Duration{}
latencies using the data in the internal/latency package.

The new latency package provides the lm.Latency(a, b) to get the
latency between two hotstuff.IDs.

We now only need to fill the Locations []string in the ReplicaOpts
to pass the set of locations around.