feat(config): first draft of cue-based configuration
The cue-based configuration allows to specify separate machines
for clients and replicas. Further, it allows to specify how many
replicas and clients should be placed on those machines. The Go
code takes care of assigning replicas and clients to machines
based on the number of replicas and clients specified in the config.
Furthermore, you may optionally specify locations where the replicas
should be placed (and corresponding latencies will be applied by
the latency framework). Finally, you may also specify treePositions
for the replicas; the treePositions refers to the positions in the
locations list, and follow a given format (root, left, right, left-left...)
Finally, you may also specify Byzantine strategies of the different
replicas in the locations list, by providnig a map from strategy
name (slow, silent) to a list of replicas that exhibit this behavior.

Please see the schema.cue for specification of constraints, and the
code for details on how to use it within the framework.