refactor(trees): remove double storage of tree position to ID map
We only need to store the IDs once; the index into the ID slice
represents that ID's position in the tree. I think it will be faster
for small trees to do this approach; we may consider to add a
benchmark to find the breaking point where it is faster to use a
map for this. If a map is needed for larger trees for speed, we
could consider to revert this change, or allocate the map only if
the tree is actually large.