refactor(latencygen): moved parsing and code generator to cli pkg
This moves the ParseCSVLatencies and GenerateGoLatencyMatrix funcs
to the cli package to make the functions available to other parts
of the codebase, if necessary. Right now, it is only used by the
latencygen tool, but the new cli.LatencyMatrix() could be used to
convert a parsed CSV file to a latency matrix at runtime instead
of using the pre-compiled code generated by GenerateGoLatencyMatrix.