fix(latencygen): include csv source file in the generated go file
Previously, we only had a latencies.csv file; we now can switch
between two different ones: aws.csv and wonderproxy.csv. This
commit fixes it so that the correct label is included in the
generated latency_matrix.go file. The latency.SourceFile variable
can be used by experiments to log the CSV source file used to run
the experiment.