refactor(kauri): this simplifies the logic of several methods++
This replaces the deprecated RegisterObserver with RegisterHandler
and the eventloop.Prioritize() option.

I replaced the Kauri reference in serviceImpl with a reference to
the eventloop instead; since we only used the eventLoop from the
Kauri struct anyway.

I replaced the time.NewTicker() (3 lines) with a simple time.Sleep,
since we should not allocate tickers for every time we call
WaitToAggregate().

I also removed the unused bool from mergeContribution; it was not
used and makes it harder to understand the code.

I replaced (k *Kauri) canMergeContributions with a function:
  func canMergeContributions(a, b hotstuff.QuorumSignature) error
that returns an error instead of a bool. This allows better error
propagation and avoids logging error messages when anyway want to
pass the error up the call stack.

I removed verifyContribution() since it was easier to process call
the block fetching and directly erroring out from the mergeContribution
method.

Finally, I revised the method:
 func (k *Kauri) mergeContribution(currentSignature hotstuff.QuorumSignature) error {
So that it returns only an error and follows the guard pattern and
thus avoids the nesting inside if-statements becomes untenable.

Please please learn the guard pattern!! It is ubiquitous across
many langauges...

if err != nil {
    return err
}
// continue with program logic (without indentation)
...
if err != nil {
    return err
}
// more program logic without indentation
...