fix(cli): added back .toml --config flag; renamed to --cue flag
This puts back the code to parse the .toml based cfgFile; this is
temporary, because we had lots of churn on this part and we had two
--config flags; a global one defined in cli/root.go and one defined
in cli/run.go. We should decide if we want to support .toml files
or not, but if we decide to not support toml, we should remove the
code and the global flag. The benefit of toml is that it works
automatically with the viper/cobra libraries.

Moreover, to avoid conflict with the global --config flag, I renamed
the one in cli/run.go to --cue so they can now both coincide.

That said, we no longer have code to process the toml file, except
for perhaps the regular flags... But all the host config must now
be processed from the cue config file.