test(hotstuffpb): reactivate most conversion tests
This reactivates most conversion tests for PartialCert, QuorumCert,
and Block. However, it remains to fix TestConvertTimeoutCertBLS12,
which currently is renamed to disabledTestConvertTimeoutCertBLS12.
See various TODO items that explains a few updates that would be
desirable to fix. I've not made any progress on this in the last
few weeks, so I'm committing it as disabled just to move on to
other aspects/tests, while thinking about this some more.