Refactor(misc)
Dependency creation no longer needs dependency set structs (except in protocol for now). This simplifies dependency creation and has shown that consensus and synchronizer are bloated, strengthening the argument that these components should be investigated and refactored. There was a lint error from previous commit which is now fixed. Finally, it should be noted that the netconfig package is only required by crypto and sender, so making this a part of globals is a good option since it only stores configuration data.