refactor: replica vs worker wiring and simplify options processing
This moves many of the module initializations out of replica.New to avoid too many config decision therein. This is of course a tradeoff, but it avoids having to duplicate options handling for the string-named modules like ConsensusRules, ByzantineStrategy, LeaderRotation, and Crypto. These now accept an empty string, which will fallthrough to the default module (chainedhotstuff, roundrobin, ecdsa).

The worker's createReplica function is now a bit too long, so it makes sense to create some wiring helper for some of the modules, e.g., the ones that need to access some ReplicaOpts field.