fix(twins): error handling in createTwinsNodes
Tried to activating TestFHSBug again. This caused createTwinsNodes
to fail silently because the error wasn't returned. This fix stops
creating nodes early if one of the nodes cannot be created.

This does not fix the underlying problem triggered by TestFHSBug;
namely that the vulnerableFHS consensusName is not known. This
needs to be fixed separately because there are conflicting changes
in PR #203.