feat(synchronizer): implement rule for signing local timeout msgs
This adds a TimeoutRuler interface and two implementations;
Simple and Aggregate. The latter is aimed at Fast-Hotstuff's aggregate QCs.
This is the first step in simplifying the synchronizer's logic by extracting
the timeout rules (or rather signing and verifying of timeout messages).
This first step only adds the LocalTimeoutRule method.