feat(synchronizer): add new VerifySyncInfo method to timeout rules
This adds the new method VerifySyncInfo to the timeout rules that
will verify various combinations of TC, QC, and AC. This finally
simplifies the advanceView logic significantly, will remaining
flexible to both aggregate and simple verification. This also adds
more tests to cover various usages that should and should not be
used.