chore: upgraded relab/iago and transitive dependencies
This upgrades the relab/iago dependency and its transitive deps.
The iago change is that we can run docker using the smaller
alpine:latest image instead of the larger ubuntu:rolling; the
latter was also causing issues for some reason. The iago and
its container dependency also allows the output stream to be
consumed by a configurable output source, such t.Output() or
os.Stdout, or a file as necessary. See relab/container and
relab/iago commit history or APIs for details.