test: add TestVerifyAggregateQCReproduceBLS12FailVerify
This attempts to reproduce the BLS12 verification failure reported
in issue #232; this is difficult to reproduce. Hence, we will need
to run it with a high -count value and -timeout=0 to avoid panicing
too early to detect the problem.

However, see issues #242 and #232, which indicate that there could
be a bug in the BLS12 dependency. Hence, it may not be worth debugging
this issue more until we have a suitable replacement for the BLS12
dependency.