fix: improve ViewStates.SyncInfo method to create before lock
This creates the new SyncInfo instance before taking the read lock
and then reading the highQC and highTC values of while holding the
ViewStates' read lock, and updating the SyncInfo accordingly. This
is really a minor thing, but it makes it clearer that we only need
the lock while reading the two fields.