refactor: simplify VerifySyncInfo in Aggregate and Simple timeout rules
This uses the new HighestView() methods to determine if the QC or
TC has the highest view, and thus if timeout should be true.

Note: Moving the initialization of view and timeout before the
signature verification should be fine, since a signature failure
will return err != nil, and users of the VerifySyncInfo must always
check for err before using any other output.