refactor(eventloop): make DelayUntil a generic function
This makes DelayUntil[T] generic on the event type to wait for.
The event type itself is still any.

The reason for making this a function is that Go cannot do generic
methods unless you specify the T on the struct, and that's not a
viable approach when we want to store multiple event types in the
same map.