feat: use client.ID and hotstuff.ID to enable metrics consistently
This selectively enables metrics only for client and replica based
on the type of ID passed to the Enable function; thus, we do not
enable client metrics logging on replicas and vice versa.

This also updates the Worker to use appropriate ID types depending
on whether it is a client or replica we are creating/starting/stopping.

This also updates the documation for the various metrics types.